==> 1.Data-Cleaning.ipynb <==
# Stage 1: Data Cleaning 
- Necessary cleaning is needed to make the words easily for the machine to process.
- Cleaning involve removing typos as well as replacing short forms into the full word.
- Tokenization and lemmatization then helps remove stopwords that are not useful for analysis


### Importing Libraries

# Importing necessary libraries
import pickle
import pandas as pd
import re
import string
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer

#Read csv file
df = pd.read_csv('AmazonMicrowave_Reviews.csv')
df

### Basic Cleaning of Data

# Basic cleaninging of dataset
def clean_text(text):
    text = re.sub('[‘’“”…]', '', text)
    text = re.sub('\n', '', text)
    text = re.sub('\xa0', '', text)
    text = re.sub(r'\[.*?\]', ' ', text)            # remove text within [ ] (' ' instead of '')
    text = re.sub(r'\<.*?\>', ' ', text)            # remove text within < > (' ' instead of '')
    text = re.sub(r'http\S+', ' ', text)            # remove website ref http
    text = re.sub(r'www\S+', ' ', text)             # remove website ref www

    text = text.replace('€', 'euros')               # replace special character with words
    text = text.replace('£', 'gbp')                 # replace special character with words
    text = text.replace('$', 'dollar')              # replace special character with words
    text = text.replace('%', 'percent')             # replace special character with words
    text = text.replace('\n', ' ')                  # remove \n in text that has it

    text = text.replace('\'', '’')                  # standardise apostrophe
    text = text.replace('&#39;', '’')               # standardise apostrophe

    text = text.replace('’d', ' would')             # remove ’ (for would, should? could? had + PP?)
    text = text.replace('’s', ' is')                # remove ’ (for is, John's + N?)
    text = text.replace('’re', ' are')              # remove ’ (for are)
    text = text.replace('’ll', ' will')             # remove ’ (for will)
    text = text.replace('’ve', ' have')             # remove ’ (for have)
    text = text.replace('’m', ' am')                # remove ’ (for am)
    text = text.replace('can’t', 'can not')         # remove ’ (for can't)
    text = text.replace('won’t', 'will not')        # remove ’ (for won't)
    text = text.replace('n’t', ' not')              # remove ’ (for don't, doesn't)

    text = text.replace('’', ' ')                   # remove apostrophe (in general)
    text = text.replace('&quot;', ' ')              # remove quotation sign (in general)

    text = text.replace('cant', 'can not')          # typo 'can't' (note that cant is a proper word)
    text = text.replace('dont', 'do not')           # typo 'don't'
    text = text.replace('weve', 'we have')          # typo 'we've'
    # text = text.replace('im', 'i am')               # typo 'i'm'

    text = re.sub(r'[^a-zA-Z0-9]', r' ', text)      # only alphanumeric left
    text = text.replace("   ", ' ')                 # remove triple empty space
    text = text.replace("  ", ' ')                  # remove double empty space
    return text

cleaned_text = lambda x: clean_text(x)

# Applying cleaning to dataset

# Product an cleaned dataset named clean_df
clean_df = pd.DataFrame(df.text.apply(cleaned_text))

# Reinsert the essential columns, rating and title

clean_df['rating'] = df.rating
clean_df['title'] = df.title

clean_df

#Convert to Pickle for later use
df.to_pickle("corpus.pkl")

#Convert cleaned data to pickle
clean_df.to_pickle('data_clean.pkl')
