==> 2.Text-Summarization.ipynb <==
# Stage 2: Text Summarization

- Summarization is needed as it extracts the essential information from the long chuck of each review.
- Using machine learning model Bert.
- Shortens the reviews while leaving it with only the main idea of the whole review.

### Importing Libraries

# Importing libraries needed for summarization
from summarizer import Summarizer,TransformerSummarizer
import pandas as pd
import re


# Import in the corpus file

df = pd.read_pickle('data_clean.pkl')

df

### Summarization

# Applying summarization on the data

bert_model = Summarizer()

df['summary'] = df['text'].apply(lambda x: bert_model(x))

df

# Removing duplicates and na data

df.drop_duplicates(subset=['summary'],inplace=True)  #dropping duplicates
df.dropna(axis=1,inplace=True)   #dropping na
df.reset_index(drop=True, inplace=True)   #Reseting index for ease of search

df

# Pickle of later use

df.to_pickle('data_clean_summarized.pkl')
