==> 4.Advanced-Analysis.ipynb <==
# Stage 4: Advanced Analysis
- Sentiment Analysis
- Text Classification

### Importing Libraries

import pandas as pd
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns

from transformers import pipeline

### Files Import

data = pd.read_pickle('data_clean_summarized.pkl')
data.reset_index(drop=True, inplace=True)

data

# from Amazon website (https://www.amazon.com/AmazonBasics-Microwave-Small-Works-Alexa/dp/B07894S727)
specs = pd.read_csv("specs.csv")
specifications = list(specs["Specifications"])                  # extract specifications headers from specs.csv

# from Mr Right Ideas website (https://www.mrright.in/ideas/appliances/microwave/microwave-oven-components-and-their-functions/)
comps = pd.read_csv("comps.csv")
components = list(dict.fromkeys(comps["Components"]))           # extract component headers from comps.csv
comps_renamed = list(dict.fromkeys(comps["Comps_renamed"]))     # extract renamed components from comps.csv

# from wiki_scraper.py
wiki_adjusted = pd.read_csv("wiki_adjusted.csv")
wiki_adjusted = list(wiki_adjusted['Keywords'])                 # extract sustainability keywords from wiki_adjusted.csv

others = ['Packaging', 'Delivery']
specifications += others

data['comps_mention'] = 'na'

### Components Mentioned in Reviews
- With Sentiment Analysis

for i in range(data.shape[0]):
    for j in comps_renamed:
        if j in data['text'][i]:
            data['comps_mention'][i] = j

data

# Create quick lambda functions to find the polarity and subjectivity of each routine
from textblob import TextBlob

count_result = []
for comps in comps_renamed:
    count = [data[data.comps_mention == comps].shape[0]]
    count_result += count

pol = lambda x: TextBlob(x).sentiment.polarity
data['polarity'] = data['summary'].apply(pol)

pol_count  = lambda x: 1  if x > 0 else -1
data['pol_count'] = data['polarity'].apply(pol_count)
display(data)

sentiment_list = []
for comps in comps_renamed:
    data_count = data[data.comps_mention == comps]
    sen_ana = data_count['pol_count'].sum()
    if sen_ana > 0:
        sen_ana = 'positive'
    else:
        sen_ana = 'negative'
    sentiment_list += [sen_ana]
    # df_chart['sentiments'] = [df_chart['labels'] == comps]

# data['pol_count'].sum()

chart = {'labels': comps_renamed,
        'values': count_result,
        'sentiments': sentiment_list}

df_chart = pd.DataFrame(chart, columns=['labels','values', 'sentiments'])

df_chart

# Data check

data_selected = data[data.comps_mention == df_chart.labels[2]] # check number accordingly
data_selected.reset_index(drop=True, inplace=True)

data_selected

### Sustainability Mentioned in Reviews
- With Sentiment Analysis

data['sus_mention'] = 'na'
wiki_adjusted = ['waste', 'eneryg', 'environmental', 'material']
for i in range(data.shape[0]):
    for j in wiki_adjusted:
        if j in data['text'][i]:
            data['sus_mention'][i] = j

data

# Create quick lambda functions to find the polarity and subjectivity of each routine
from textblob import TextBlob

count_result = []
for sus in wiki_adjusted:
    count = [data[data.sus_mention == sus].shape[0]]
    count_result += count

pol = lambda x: TextBlob(x).sentiment.polarity
data['polarity'] = data['summary'].apply(pol)

pol_count  = lambda x: 1  if x > 0 else -1
data['pol_count'] = data['polarity'].apply(pol_count)
display(data)

sentiment_list = []
for sus in wiki_adjusted:
    data_count = data[data.sus_mention == sus]
    sen_ana = data_count['pol_count'].sum()
    if sen_ana > 0:
        sen_ana = 'positive'
    else:
        sen_ana = 'negative'
    sentiment_list += [sen_ana]
    # df_chart['sentiments'] = [df_chart['labels'] == comps]

# data['pol_count'].sum()

chart = {'labels': wiki_adjusted,
        'values': count_result,
        'sentiments': sentiment_list}

df_chart = pd.DataFrame(chart, columns=['labels','values', 'sentiments'])

df_chart

# Data check

data_selected = data[data.comps_mention == df_chart.labels[2]] # check number accordingly
data_selected.reset_index(drop=True, inplace=True)

data_selected

## Zero-Shot-Classification

classifier = pipeline("zero-shot-classification", device=0)

### Specification Categorization
- With Sentiment Analysis 

display(specifications)
#Applying it into the dataframe base on original review
data['category'] = data['text'].apply(lambda x: classifier(x, candidate_labels=specifications))

data

#Replacing the category with the top matched category
for i in tqdm(range(data.shape[0])):
    if data['category'][i] not in specifications:
        data['category'][i] = data['category'][i].get('labels')[0]

data_cat = data.copy()

candidate_results = list(0 for i in range(len(specifications)))


def count(category):
    for i in range(len(specifications)):
        if category == specifications[i]:
            candidate_results[i] = candidate_results[i] + 1

# Run the counting function
data_cat['category'].apply(lambda x: count(x))

candidate_results


sentiment_list_func = []
for func in specifications:
    data_count = data[data.category == func]
    sen_ana = data_count['pol_count'].sum()
    if sen_ana > 0:
        sen_ana = 'positive'
    else:
        sen_ana = 'negative'
    sentiment_list_func += [sen_ana]
    # df_chart['sentiments'] = [df_chart['labels'] == comps]

# data['pol_count'].sum()

chart = {'labels': specifications,
        'values': candidate_results,
        'sentiments': sentiment_list_func}

df_chart = pd.DataFrame(chart, columns=['labels','values', 'sentiments'])

df_chart

### Sustainability Categorization 
- Sentiment Analysis included

#Categorying in terms of sustainability

#Applying it into the dataframe base on original review
data['sus_category'] = data['text'].apply(lambda x: classifier(x, candidate_labels=wiki_adjusted))

data

#Replacing the category with the top matched category
for i in tqdm(range(data.shape[0])):
    if data['sus_category'][i] not in wiki_adjusted:
        data['sus_category'][i] = data['sus_category'][i].get('labels')[0]

data_cat = data.copy()

candidate_results = list(0 for i in range(len(wiki_adjusted)))


def count(category):
    for i in range(len(wiki_adjusted)):
        if category == wiki_adjusted[i]:
            candidate_results[i] = candidate_results[i] + 1

# Run the counting function
data_cat['sus_category'].apply(lambda x: count(x))

candidate_results


sentiment_list_func = []
for func in wiki_adjusted:
    data_count = data[data.category == func]
    sen_ana = data_count['pol_count'].sum()
    if sen_ana > 0:
        sen_ana = 'positive'
    else:
        sen_ana = 'negative'
    sentiment_list_func += [sen_ana]
    # df_chart['sentiments'] = [df_chart['labels'] == comps]

# data['pol_count'].sum()

chart = {'labels': wiki_adjusted,
        'values': candidate_results,
        'sentiments': sentiment_list_func}

df_chart = pd.DataFrame(chart, columns=['labels','values', 'sentiments'])

df_chart

data.to_pickle('data_analysed.pkl')

data
